It is unclear now how this test can be implemented on 4.0 after introduction of READ CONSISTENCY with default value ReadConsistency = 1. According to doc: ===     If ReadConsistency set to 1 (by default) engine ignores     [NO] RECORD VERSION flags and makes all read-committed     transactions READ COMMITTED READ CONSISTENCY. ===
Also, doc\README.read_consistency.md  says that:
"Old read-committed isolation modes (**RECORD VERSION** and **NO RECORD VERSION**) are still 
allowed but considered as legacy and not recommended to use."

This mean that one can NOT to check issues of this ticket under 4.0 using default (and recommended)
value of config parameter 'ReadConsistency'.
For that reason it was decided to make new EMPTY section of test for 4.0.